Check that beam.specs satisfies all necessary conditions

Check that beam.specs satisfies all necessary conditions

Usage

check_beam_specs(beam.specs, mtx.names)

Arguments

beam.specs

A data.frame with column name, mtx, and mdl

mtx.names

A vector with the names of the data matrices (beam.data$mtx.data)

Value

A data.frame of beam.specs if all conditions satisfied, otherwise throws an error

Examples

data(beam_dat)
data(beam_specs)
test_specs <- check_beam_specs(beam_specs, names(beam_dat$mtx.data))