Extract beam stats for a specific set

Extract beam stats for a specific set

Usage

extract_beam_stats(beam.stats, set.id)

Arguments

beam.stats

A beam.stats object, which is a list with beam.stats (the association matrices), the beam.specs, and the beam.data

set.id

A character of a set id name (an entry in in beam.data$set.data$set.id)

Value

A matrix with with estimated associations for each endpoint and each omic feature linked to the set

Examples

# \donttest{
test.beam.data <- prep_beam_data(main.data=clinf, mtx.data=omicdat,
                                 mtx.anns=omicann, set.data=setdat,
                                 set.anns=NULL, n.boot=10, seed=123)
#> Checking inputs: Fri Jul 12 18:37:42 2024
#>   Checking that each element of mtx.data is a matrix: Fri Jul 12 18:37:42 2024
#>   Checking that each element of mtx.anns is a data.frame: Fri Jul 12 18:37:42 2024
#> Aligning main.data with each mtx.data: Fri Jul 12 18:37:42 2024
#>   Working on mtx.data Lesion (1 of 2): Fri Jul 12 18:37:42 2024
#>   Working on mtx.data RNA (2 of 2): Fri Jul 12 18:37:42 2024
#> Warning: Some ids not matched; returning NAs for those.
#>   Working on mtx.anns: Fri Jul 12 18:37:42 2024
#>   Matching matrix 1 with annotations: Fri Jul 12 18:37:42 2024
#>   Matching matrix 2 with annotations: Fri Jul 12 18:37:42 2024
#>   Checking set.data: Fri Jul 12 18:37:42 2024
#>     Ordering and indexing set.data: Fri Jul 12 18:37:42 2024
#>     Checking section 1 of 40 of set.data: Fri Jul 12 18:37:42 2024
#> Generating bootstrap index matrix: Fri Jul 12 18:37:42 2024
#> Packaging and returning result: Fri Jul 12 18:37:42 2024
specs <- prep_beam_specs(beam.data=test.beam.data, endpts=c("MRD29", "EFS"),
                         firth=TRUE)
#> MRD29 is continuous, fitting lm
#> EFS is survival endpoint, fitting coxphf2
test.beam.stats <- compute_beam_stats(beam.data=test.beam.data,
                                      beam.specs=specs)
#> Working on analysis 1 of 4 : Fri Jul 12 18:37:42 2024
#> Lesionlm(MRD29~mtx.row,data=main.data,model=T)
#>   Working on bootstrap 0 of 11 : Fri Jul 12 18:37:42 2024
#>   Working on bootstrap 10 of 11 : Fri Jul 12 18:37:42 2024
#> Working on analysis 2 of 4 : Fri Jul 12 18:37:42 2024
#> RNAlm(MRD29~mtx.row,data=main.data,model=T)
#>   Working on bootstrap 0 of 11 : Fri Jul 12 18:37:42 2024
#>   Working on bootstrap 10 of 11 : Fri Jul 12 18:37:42 2024
#> Working on analysis 3 of 4 : Fri Jul 12 18:37:42 2024
#> Lesioncoxphf2(EFS~mtx.row,data=main.data,model=T)
#>   Working on bootstrap 0 of 11 : Fri Jul 12 18:37:42 2024
#>   Working on bootstrap 10 of 11 : Fri Jul 12 18:37:57 2024
#> Working on analysis 4 of 4 : Fri Jul 12 18:37:59 2024
#> RNAcoxphf2(EFS~mtx.row,data=main.data,model=T)
#>   Working on bootstrap 0 of 11 : Fri Jul 12 18:37:59 2024
#>   Working on bootstrap 10 of 11 : Fri Jul 12 18:38:16 2024
test.stats <- extract_beam_stats(test.beam.stats, set.id="ENSG00000099810")
# }
data(beam_stats)
test.stats <- extract_beam_stats(beam_stats, set.id="ENSG00000099810")