Subset beam.stats Result

Filter the beam.stats object from compute_beam_stats with various filtering criteria. Default is to filter to top 50 sets with smallest q-value. At least one filtering criteria must be specified. Can also use intersection or union of multiple criteria.

Usage

subset_beam_result(
  beam.result,
  beam.set.pvals = NULL,
  beam.feat.pvals = NULL,
  mtx.rows = NULL,
  set.ids = NULL,
  endpts = NULL,
  omics = NULL,
  p.limit = NULL,
  q.limit = NULL,
  p.feat.limit = NULL,
  q.feat.limit = NULL,
  intersect = TRUE,
  recalc = FALSE
)

Arguments

beam.result

A beam.stats object from compute_beam_stats

beam.set.pvals

A list containing BEAMR set p-values from compute_set_pvalues; required if p.limit or q.limit are specified.

beam.feat.pvals

A list containing feature-level p-values from compute_feature_pvalues; required if p.feat.limit or q.feat.limit are specified.

mtx.rows

A list of vectors of feature names corresponding to row.id in set.data. List names correspond to mtx.id in set.data. If specified, filter to all sets containing at least one of these features.

set.ids

A character vector of set.ids. If specified, filter to these sets.

endpts

A character vector of endpoint names. If specified, filter to sets that correspond to these endpoints.

omics

A character vector of omics names. If specified, fitler to sets that correspond to these omics.

p.limit

A numeric value. If specified, determine mtx.rows that are below this threshold if p<1 or top p sets if p>1.

q.limit

A numeric value. If specified, determine mtx.rows that are below this threshold if q <1 or top q sets if q>1.

p.feat.limit

A numeric value. If specified, determine mtx.rows that are below this threshold if p.feat<1 or top p.feat sets if p.feat>1 (feature p-values).

q.feat.limit

A numeric value. If specified, determine mtx.rows that are below this threshold if q.feat<1 or top q.feat sets if q.feat>1.

intersect

A logical value. Default is TRUE. If TRUE, use intersection of all specified criteria. If FALSE use union of all specified criteria.

recalc

A logical value. Default is FALSE. If TRUE, recalculate p-values. If FALSE use original set p-values..

Value

A list with filtered beam.stats object, updated beam.set.pvals, and filtered beam.feat.pvals.

Examples

# \donttest{
test.beam.data <- prep_beam_data(main.data=clinf, mtx.data=omicdat,
                                 mtx.anns=omicann, set.data=setdat,
                                 set.anns=NULL, n.boot=10, seed=123)
#> Checking inputs: Fri Jul 12 18:42:02 2024
#>   Checking that each element of mtx.data is a matrix: Fri Jul 12 18:42:02 2024
#>   Checking that each element of mtx.anns is a data.frame: Fri Jul 12 18:42:02 2024
#> Aligning main.data with each mtx.data: Fri Jul 12 18:42:02 2024
#>   Working on mtx.data Lesion (1 of 2): Fri Jul 12 18:42:02 2024
#>   Working on mtx.data RNA (2 of 2): Fri Jul 12 18:42:02 2024
#> Warning: Some ids not matched; returning NAs for those.
#>   Working on mtx.anns: Fri Jul 12 18:42:02 2024
#>   Matching matrix 1 with annotations: Fri Jul 12 18:42:02 2024
#>   Matching matrix 2 with annotations: Fri Jul 12 18:42:02 2024
#>   Checking set.data: Fri Jul 12 18:42:02 2024
#>     Ordering and indexing set.data: Fri Jul 12 18:42:02 2024
#>     Checking section 1 of 40 of set.data: Fri Jul 12 18:42:02 2024
#> Generating bootstrap index matrix: Fri Jul 12 18:42:02 2024
#> Packaging and returning result: Fri Jul 12 18:42:02 2024
specs <- prep_beam_specs(beam.data=test.beam.data, endpts=c("MRD29", "EFS", "OS"),
                         firth=TRUE)
#> MRD29 is continuous, fitting lm
#> EFS is survival endpoint, fitting coxphf2
#> OS is survival endpoint, fitting coxphf2
test.beam.stats <- compute_beam_stats(beam.data=test.beam.data, beam.specs=specs)
#> Working on analysis 1 of 6 : Fri Jul 12 18:42:02 2024
#> Lesionlm(MRD29~mtx.row,data=main.data,model=T)
#>   Working on bootstrap 0 of 11 : Fri Jul 12 18:42:02 2024
#>   Working on bootstrap 10 of 11 : Fri Jul 12 18:42:02 2024
#> Working on analysis 2 of 6 : Fri Jul 12 18:42:02 2024
#> RNAlm(MRD29~mtx.row,data=main.data,model=T)
#>   Working on bootstrap 0 of 11 : Fri Jul 12 18:42:02 2024
#>   Working on bootstrap 10 of 11 : Fri Jul 12 18:42:02 2024
#> Working on analysis 3 of 6 : Fri Jul 12 18:42:02 2024
#> Lesioncoxphf2(EFS~mtx.row,data=main.data,model=T)
#>   Working on bootstrap 0 of 11 : Fri Jul 12 18:42:02 2024
#>   Working on bootstrap 10 of 11 : Fri Jul 12 18:42:17 2024
#> Working on analysis 4 of 6 : Fri Jul 12 18:42:19 2024
#> RNAcoxphf2(EFS~mtx.row,data=main.data,model=T)
#>   Working on bootstrap 0 of 11 : Fri Jul 12 18:42:19 2024
#>   Working on bootstrap 10 of 11 : Fri Jul 12 18:42:37 2024
#> Working on analysis 5 of 6 : Fri Jul 12 18:42:39 2024
#> Lesioncoxphf2(OS~mtx.row,data=main.data,model=T)
#>   Working on bootstrap 0 of 11 : Fri Jul 12 18:42:39 2024
#>   Working on bootstrap 10 of 11 : Fri Jul 12 18:42:54 2024
#> Working on analysis 6 of 6 : Fri Jul 12 18:42:55 2024
#> RNAcoxphf2(OS~mtx.row,data=main.data,model=T)
#>   Working on bootstrap 0 of 11 : Fri Jul 12 18:42:55 2024
#>   Working on bootstrap 10 of 11 : Fri Jul 12 18:43:12 2024
test.pvals <- compute_set_pvalues(beam.stats=test.beam.stats)
#> Preparing bootstrap results for calculating feature set p-values: Fri Jul 12 18:43:14 2024
#> Finding stats for each data matrix:Fri Jul 12 18:43:14 2024
#>   Finding stats for data matrix Lesion: Fri Jul 12 18:43:14 2024
#>    Finding features with with Lesion.MRD29 stats: Fri Jul 12 18:43:14 2024
#>    Finding features with with Lesion.EFS stats: Fri Jul 12 18:43:14 2024
#>    Finding features with with Lesion.OS stats: Fri Jul 12 18:43:14 2024
#>   Finding stats for data matrix RNA: Fri Jul 12 18:43:14 2024
#>    Finding features with with RNA.MRD29 stats: Fri Jul 12 18:43:14 2024
#>    Finding features with with RNA.EFS stats: Fri Jul 12 18:43:14 2024
#>    Finding features with with RNA.OS stats: Fri Jul 12 18:43:14 2024
#> Found 120 rows of stats: Fri Jul 12 18:43:14 2024
#> Merging stats with feature-sets: Fri Jul 12 18:43:14 2024
#> Merged feature-set stat rows: 120
#> Ordering and indexing feature sets: Fri Jul 12 18:43:14 2024
#>   Cleaning up beam.stat matrices:Fri Jul 12 18:43:14 2024
#>    Working on matrix 1 of 6: Fri Jul 12 18:43:14 2024
#>    Working on matrix 2 of 6: Fri Jul 12 18:43:14 2024
#>    Working on matrix 3 of 6: Fri Jul 12 18:43:14 2024
#>    Working on matrix 4 of 6: Fri Jul 12 18:43:14 2024
#>    Working on matrix 5 of 6: Fri Jul 12 18:43:14 2024
#>    Working on matrix 6 of 6: Fri Jul 12 18:43:14 2024
#> Computing p-value for feature set 1 of 34: Fri Jul 12 18:43:14 2024
#> 13ENSG00000081760
#> Computing p-value for feature set 26 of 34: Fri Jul 12 18:43:14 2024
#> 9496ENSG00000229835
#> Finished computing p-values at: Fri Jul 12 18:43:14 2024
#> Minimum q-value is 0.106439437831269
#> Creating set p-value data frame.
#> Done creating data frame.
#> Creating list for output.
test.feat.pvals <- compute_feature_pvalues(beam.stats=test.beam.stats)
#> Computing feature p-values for stat matrix 1 of 6: Fri Jul 12 18:43:14 2024
#>   This matrix has 20 features.
#> Computing feature p-values for stat matrix 2 of 6: Fri Jul 12 18:43:14 2024
#>   This matrix has 20 features.
#> Computing feature p-values for stat matrix 3 of 6: Fri Jul 12 18:43:14 2024
#>   This matrix has 20 features.
#> Computing feature p-values for stat matrix 4 of 6: Fri Jul 12 18:43:14 2024
#>   This matrix has 20 features.
#> Computing feature p-values for stat matrix 5 of 6: Fri Jul 12 18:43:14 2024
#>   This matrix has 20 features.
#> Computing feature p-values for stat matrix 6 of 6: Fri Jul 12 18:43:14 2024
#>   This matrix has 20 features.
filt.beam.stats <- subset_beam_result(test.beam.stats, test.pvals,
                                      test.feat.pvals, endpts=c("EFS", "OS"),
                                      q.limit=10, intersect=TRUE,
                                      recalc=FALSE)
#>   Working on mtx.anns: Fri Jul 12 18:43:14 2024
#>   Matching matrix 1 with annotations: Fri Jul 12 18:43:14 2024
#>   Matching matrix 2 with annotations: Fri Jul 12 18:43:14 2024
#> Joining with `by = join_by(set.id, mtx.id, row.id)`
#>   Working on mtx.anns: Fri Jul 12 18:43:14 2024
#>   Matching matrix 1 with annotations: Fri Jul 12 18:43:14 2024
#>   Matching matrix 2 with annotations: Fri Jul 12 18:43:14 2024
#> Computing feature p-values for stat matrix 1 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 6 features.
#> Computing feature p-values for stat matrix 2 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 7 features.
#> Computing feature p-values for stat matrix 3 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 6 features.
#> Computing feature p-values for stat matrix 4 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 7 features.
#> Computing feature p-values for stat matrix 1 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 6 features.
#> Computing feature p-values for stat matrix 2 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 7 features.
#> Computing feature p-values for stat matrix 3 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 6 features.
#> Computing feature p-values for stat matrix 4 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 7 features.
# }
data(beam_stats)
test.pvals <- compute_set_pvalues(beam.stats=beam_stats)
#> Preparing bootstrap results for calculating feature set p-values: Fri Jul 12 18:43:14 2024
#> Finding stats for each data matrix:Fri Jul 12 18:43:14 2024
#>   Finding stats for data matrix Lesion: Fri Jul 12 18:43:14 2024
#>    Finding features with with Lesion.MRD29 stats: Fri Jul 12 18:43:14 2024
#>    Finding features with with Lesion.EFS stats: Fri Jul 12 18:43:14 2024
#>    Finding features with with Lesion.OS stats: Fri Jul 12 18:43:14 2024
#>   Finding stats for data matrix RNA: Fri Jul 12 18:43:14 2024
#>    Finding features with with RNA.MRD29 stats: Fri Jul 12 18:43:14 2024
#>    Finding features with with RNA.EFS stats: Fri Jul 12 18:43:14 2024
#>    Finding features with with RNA.OS stats: Fri Jul 12 18:43:14 2024
#> Found 120 rows of stats: Fri Jul 12 18:43:14 2024
#> Merging stats with feature-sets: Fri Jul 12 18:43:14 2024
#> Merged feature-set stat rows: 120
#> Ordering and indexing feature sets: Fri Jul 12 18:43:14 2024
#>   Cleaning up beam.stat matrices:Fri Jul 12 18:43:14 2024
#>    Working on matrix 1 of 6: Fri Jul 12 18:43:14 2024
#>    Working on matrix 2 of 6: Fri Jul 12 18:43:14 2024
#>    Working on matrix 3 of 6: Fri Jul 12 18:43:14 2024
#>    Working on matrix 4 of 6: Fri Jul 12 18:43:14 2024
#>    Working on matrix 5 of 6: Fri Jul 12 18:43:14 2024
#>    Working on matrix 6 of 6: Fri Jul 12 18:43:14 2024
#> Computing p-value for feature set 1 of 34: Fri Jul 12 18:43:14 2024
#> 13ENSG00000081760
#> Computing p-value for feature set 26 of 34: Fri Jul 12 18:43:14 2024
#> 9496ENSG00000229835
#> Finished computing p-values at: Fri Jul 12 18:43:14 2024
#> Minimum q-value is 0.106439437831269
#> Creating set p-value data frame.
#> Done creating data frame.
#> Creating list for output.
test.feat.pvals <- compute_feature_pvalues(beam.stats=beam_stats)
#> Computing feature p-values for stat matrix 1 of 6: Fri Jul 12 18:43:14 2024
#>   This matrix has 20 features.
#> Computing feature p-values for stat matrix 2 of 6: Fri Jul 12 18:43:14 2024
#>   This matrix has 20 features.
#> Computing feature p-values for stat matrix 3 of 6: Fri Jul 12 18:43:14 2024
#>   This matrix has 20 features.
#> Computing feature p-values for stat matrix 4 of 6: Fri Jul 12 18:43:14 2024
#>   This matrix has 20 features.
#> Computing feature p-values for stat matrix 5 of 6: Fri Jul 12 18:43:14 2024
#>   This matrix has 20 features.
#> Computing feature p-values for stat matrix 6 of 6: Fri Jul 12 18:43:14 2024
#>   This matrix has 20 features.
filt.beam.stats <- subset_beam_result(beam_stats, test.pvals, test.feat.pvals,
                                      endpts=c("EFS","OS"), q.limit=10, intersect=TRUE,
                                      recalc=FALSE)
#>   Working on mtx.anns: Fri Jul 12 18:43:14 2024
#>   Matching matrix 1 with annotations: Fri Jul 12 18:43:14 2024
#>   Matching matrix 2 with annotations: Fri Jul 12 18:43:14 2024
#> Joining with `by = join_by(set.id, mtx.id, row.id)`
#>   Working on mtx.anns: Fri Jul 12 18:43:14 2024
#>   Matching matrix 1 with annotations: Fri Jul 12 18:43:14 2024
#>   Matching matrix 2 with annotations: Fri Jul 12 18:43:14 2024
#> Computing feature p-values for stat matrix 1 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 6 features.
#> Computing feature p-values for stat matrix 2 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 7 features.
#> Computing feature p-values for stat matrix 3 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 6 features.
#> Computing feature p-values for stat matrix 4 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 7 features.
#> Computing feature p-values for stat matrix 1 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 6 features.
#> Computing feature p-values for stat matrix 2 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 7 features.
#> Computing feature p-values for stat matrix 3 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 6 features.
#> Computing feature p-values for stat matrix 4 of 4: Fri Jul 12 18:43:14 2024
#>   This matrix has 7 features.